Update README.md
The objective of the project was to analyse the trend of development indicators, Household Consumption, Capital Stock, Merchandise Exports, Merchandise Imports, Human Capital and Population vs GDP over a period, of all countries. The data was sourced from https://fred.stlouisfed.org/categories/33402/downloaddata. The result of the analysis was in the form of an animated graph which showed the trend. Language/Tools: Python3, Numpy, Tkinter, Matplotlib.